requirements.txt
django
django-extensions
django-crispy-forms
django-bootstrap4
djangorestframework
django-autocomplete-light>=2.0.0pre
django-tables2
django-filter
Pillow
argon2-cffi
PyPDF2
pip
django-xhtml2pdf
git+https://github.com/me2d09/pinax-notifications.git
backports-datetime-fromisoformat